pyproject.toml
[project]
name = "north-indian-food-suggester"
version = "0.1.0"
description = "Multi-agent AI system for North Indian food menu suggestions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.3.0",
    "mcp>=0.9.0",
    "pydantic>=2.0.0",
    "fastmcp>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
food-suggester = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
